- CHG: Code simplification/optimization. - FIX: Fixed an issue with the trials/compound stored procedure where a call would ignore the specified dataset ids and just export across datasets. - CHG: Reduced the number of cores used by the build process to half of the available processors. -

+++ src/jhi/germinate/server/util/ApplicationListener.java
@@ -113,8 +113,11 @@ public class ApplicationListener implements ServletContextListener
 		{
 			File[] files = file.listFiles((dir, name) -> name.matches(context + "-log-.*\\.txt.*"));
 
-			for (File f : files)
-				f.delete();
+			if (files != null)
+			{
+				for (File f : files)
+					f.delete();
+			}
 		}
 
 		// Now deregister JDBC drivers in this context's ClassLoader: Get the webapp's ClassLoader
